issue #1452 - reduce severity of supplemental issues in conformsTo
Previously, we added supplemental issues with a Severity of ERROR when
the conformance check failed.
However, we also return false in these cases and so it should be the
caller's responsibility to determine if that will result in an error or
not.
Instead, the detail in these supplemental issues is really just extra
information about the failed conformance check and thus these now have a
severity of "information".

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>