issue #1309 - update carin-bb ig and examples
This commit updates us to version 0.1.3 (pulled 9/9/2020 in prep for
FHIR connecathon 25).
The carin-bb ig is still a work-in-progress and there were a number of
issues with the examples, mostly coding-related.

Additionally, I had to fix one of the constraints that was included in
the Inpatient and Outpatient Institutional EoB profiles.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>