issue #1309 - fixup EOB-inst-careTeam invariants
Previously, I had patched the `EOB-inst-careTeam-practitioner`
constraint, changing the right-hand of the `in` operator from `or` to
`union (|)` to make the examples pass validation.

However, it turns out that `EOB-inst-careTeam-practitioner` *and*
`EOB-inst-careTeam-organization` are broken in a couple more ways.
1. They are specified at the EoB.careTeam level (which sets that as the
context), but the expression includes `careTeam` in the paths.  I fixed
by moving these two invariants to the "top level" of the structure (but
did not fix the NarrativeText at this time).
2. They ignore repeating elements.  I fixed this by adding where clauses
to select just the nodes that are applicable for the given conditions.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>