issue #1508 - Initial delivery of fhir-meter project
Detail: This is the initial delivery of the fhir-meter project
containing a readme and a JMeter test plan.  The test plan is for
searching a FHIR server with preloaded resources where you can get a
list of resource ID's into a CSV file before you run the test.  This is
meant to add load to the server through JMeter load testing.


Signed-off-by: Dag O. Roppe <roppe@us.ibm.com>