issue #1598 - Change location of copyright and additional simple search
Detail:  The original location of copyright caused the copyright to
disappear from the file as soon as it was saved once.  Now the copyright
is in a statement comment, so part of the first 2 statements in the
JMeter Test Plan.
Added additional simple searches on patient for different resource
types.


Signed-off-by: Dag O. Roppe <roppe@us.ibm.com>