Add reference target annotations, javadoc, and validation
Previously, we only had these for fields of type reference with a max
cardinality of 1. Now we have these for both repeating fields and for
choice types which include references as well.

I think there are many more cases we're missing wrt validation of choice
types and repeating fields of other types, but I was focused mainly on
ReferenceTargets for now (to address
https://github.com/IBM/FHIR/issues/1536 )

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>