issue #1536 - generate and commit updated examples
The new checks found a couple more spec examples with invalid content,
so the expectation for those is updated in the corresponding index rows.

I also noticed that the DataAbsent examples have significant updates as
well...presumably from a previous round of edits that were made but
which were never used to generated updated examples yet.

I also snuck in a minor update in CompleteMockDataCreator (for
references with no target profiles).


Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>