Private Memory on Db2 results in -930 errors at load #1527
This error is only on the Db2 and isolated to the Db2 stored procedure - add_any_resource.

Scenario
- Load a Bundle with multiple if not hundreds of resources in a simultaneous way within a single Connection
- Note, the Bundle can be set as Transaction or Batch.
- Db2 with 8GB of memory without INSTANCE_MEMORY=Automatic
- The simultaneous load with ~10 to 20 threads.
- The result is *SQL0930N There is not enough storage available to process the statement*.

In the underlying Memory Allocation, the 2GB size cause a block of memory to be reserved, where it accumulates. So if a new large request comes up requiring substantial chunk of memory, this chunk can only be freshly acquired from so far unused range of addresses, or it may fail.  If the p_payload is lowered to 10M, then the code operates as expected and the same workload can function with 200 threads. Note, it's dependent on the number of CALL to add_any_resource.

This PR is opened to discuss in more detail on an approach.  I'd like to lower the 2GB to 10M.  If needed we can branch to a LARGE size Stored Procedure if the file is larger than 10M.

Reference: https://www.ibm.com/support/pages/db2-content-management-application-may-receive-sql0930n
Signed-off-by: Paul Bastide <pbastide@us.ibm.com>