issue #1673 - change default ordering
Previously, we used a default sort of RESOURCE_ID ASC (the row identifier of the resource version being returned).
While investigating #1673, I noticed that removing this sort took this query from 10 seconds down to 3-4 seconds.

However, having *some* sort order is important, so that we get consistent page results.
Upon further inspection, I found that changing the default sort to LOGICAL_RESOURCE_ID (the row identifier of the logical resource row being returned)
yielded a similar performance boost and so this is the proposed change.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>