issue #1881 - add logic to allow reading of provenance resources
The change is specific to `afterSearch` and only allows Provenance
resources to be returned as part of a search which also returns one or
more target (so that we can ensure the user has access to at least one
target resource from the Provenance)

Additionally, I beefed up the javadoc for
ReferenceUtil.createReferenceValueFrom and ReferenceValue from
fhir-search since I had to study them to use them.

Finally, I added a variant of TestUtil.getMinimalResource which doesn't
force the Format argument (which is almost never relevant).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>