issue #1881 - support direct access to Provenance resources
Previously, we only handled Provenance access in the case of search.
Direct access (i.e. read/vread/history) to Provenance was not allowed
unless it was in the patient compartment.
Now, in the afterRead/afterVread/afterHistory methods, we will check for
Provenance resources and grant access if the Provenance targets one or
more resources that the user has access to.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>