issue #1907 - strip the query from originalRequestUri before parsing
Previously I updated our UriBuilder for this case (and introduced tests
to ensure it is properly functioning). However, I missed the fact that
FHIRREstServletFilter parses the originalRequestUri header value into a
URI and, when this fails, it simply reverts to the actual URL of the
request.

Also performed some minor cleanup.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>