set x_RESOURCE_TOKEN_REFS to V0009
We missed this as part of #1996, so anyone that has applied these updates to a DB since that change now has the wrong
version of that object.  Since that change was never released, I think we're OK just to fix it now (and help anyone
riding the bleeding edge to manually fix their VERSION_HISTORY table).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>