issue #2040 - address review comments
Remove the dataStream.close() from BulkDataUtils.multiPartUpload and add
javadoc to make it clear this isn't closed. Also some formatting/cleanup
in classes which call this (while I was making sure none of them need a
to be closed).

Additionally, I updated the ImportTransientUserData.Builder builder
signatures to return ImportTransientUserData builders and objects
instead of ImportCheckPointData; accomplished this by following the same
pattern we use in fhir-model.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>