issue #2040 - additional changes to config properties and legacy export
I think that we inadvertently changed the S3Provider.writeResources
logic as part of a late change in #1999 (the one where we set
`chunkData.isFinishCurrentUpload` when getPageNum > lastPageNum).
Specifically, in the NDJSON case, we still want to start writing parts
as soon as we have crossed the partUploadTriggerSize. This attempts to
fix that and removes some now-unnecessary `or` expressions from
S3Provider.

I also made a number of config-related changes (some of which are
probably more opinion-based than strictly needed):
* Reconciled the Fast impl's `partUploadTriggerSize` with the legacy
impl's `minPartSize` (by adopting the fast impl's name)
* Updated the maxFileSizeThreshold and minPartSize params to
`objectSizeThreshold` and `objectResourceCountThreshold` respectively
* Added "MB" after the size names to make it clear what the units are
without needing to look at the docs
* Changed ExportCheckpointAlgorithm to handle threshold values of 0 like
"no threshold" (formerly these were -1 but that is now impossible to set
due to the fact we multiply the config values in our getters)
* Fixed the javadoc in ConfigurationAdapter accordingly

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>