issue #2094 - add support for 'or' within inclusion criteria
1. Introduces SearchUtil.buildInclusionCriteria() for building the
inclusion criteria parameter with multiple compartmentId values (maybe
this should go in CompartmentUtil instead?)
2. Updates JDBCQueryBuilder.processInclusionCriteria to use all of the
inclusionCriteria values; previously it used `=` against the first value
but now it uses `IN (val1,val2,val3,...)` instead.

And update 3 places to use it:
1. A new test in AbstractSearchCompartmentTest
2. Patient export (PatientResourceHandler.fillChunkDataBuffer()
    This should be much simpler now and also much more efficient.
Instead of searching with a value of `id1,id2,id3,...` on each and every
inclusion criteria, we can do a single query with a whole page of
patient ids.
3. AuthzPolicyEnforcementPersistenceInterceptor.beforeSearch()
    This lays the groundwork for supporting `user` scopes. Instead of
converting the request to a compartment search for a single patient, we
can 'or' between the full set of patient_id values included in the
access token.
Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>