fix for Patient export ChunkReader
I was accidentally setting resourceType from
`ctx.getFhirResourceType()`, but this job parameter is actually a
comma-delimited list of resource types to export.  What I wanted instead
is the *partition* resourceType.

I fixed the mistake and then renamed `fhirResourceType` to
`fhirResourceTypes` in all the places (other than the job parameter
itself), and added some javadoc to BulkDataContext to hopefully avoid
similar confusion in the future.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>