issue #2145 - set the StandardsStatus appropriately
I had used the wiki page at
https://confluence.hl7.org/display/FHIR/FHIR+Maturity+Model to determine
what was draft/trial/normative, but in the StructureDefinition resources
there was a different field for this.

Net: my prior commit used "Trial Use" for FFM5, but the normative
resources actually have a level of 5 as well...we just needed to pull
the status from the other extension.

1. generate a code subtype class for StandardsStatus
2. change MaturityLeve(value) to Maturity(level, status)
3. regenerate the model

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>