issues #2241 and #2057 - load operation from registry and skip warning
With this pull request, the spec-defined OperationDefinitions are now
pulled from the registry instead of having them copied to each
individual module.

Additionally, because so many of these spec-defined OperationDefinitions
violate opd-0 (`name.matches('[A-Z]([A-Za-z0-9_]){0,254}')`), I toned
down the logging so that only validation errors are logged by default
(and not validation warnings).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>