AuthzPolicyEnforcement changes
1. getPatientRefValue now uses ReferenceUtil to obtain the id of the
target Patient being referenced. This technique matches what we do
during the extraction of the compartment membership param and is more
robust to special reference values like versioned references or absolute
references that match the base server url.

2. getPatientIdFromToken will now split the patient_id claim on spaces.
This sets us up for scenarios where a single token can grant access to
multiple patient records. Additionally, I removed the log statement
about expecting patient_id to be a string and not a list.

3. Patient search is now allowed. Because we convert all searches into
compartment searches, there is no longer a risk of exposing whether a
given Patient resource by a given id exists or not. Therefore, we don't
need to prevent such searches any longer. That said, because a Patient
is not in its own compartment (unless it has a corresponding
Patient.link), Patient search behavior will likely to be confusing to
clients, operators may want to continue to disallow Patient search from
the REST layer in select deployments.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>