fixup
1. revert changes to PathAwareVisitor and ResourceFingerPrintVisitor
I had made some changes to support re-use of the
ResourceFingerPrintVisitor, but this object should be light to construct
and so I've reverted those changes and used it as it was previously
designed.

2. changed onlyIfModified from a Boolean to a boolean; testing confirmed
that any non-true values will get properly mapped to false, which is
just what we want.

3. fixed parameter ordering issue in FHIRRestHelper...the new
skippableUpdate param should always come before the doValidation param
(because doValidation is only included in the overloaded variant of
update).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>