issues #2281 and #864 - batch/transaction bundle processing changes
1. FHIRRestHelper.validateBundle now returns a sparse array (modeled as
a map) of bundle response entries with warnings. It used to construct an
entire response bundle.

2. FHIRRestHelper.processEntriesByMethod replaces
FHIRRestHelper.processEntriesForMethod and iterates the request entries
a single time, instead of once-per-method.

3. Entries are collected into an array instead of "modifying" the
responseBundle with each one (which forced lots of needless copying
since our model is now immutable).

4. Stop duplicating custom operation results in the
`Bundle.entry.response.outcome` field of the response. We already copy
the operation's result in the `Bundle.entry.resource` and that should be
sufficient.

I also made significant formatting changes in FHIRRestHelper.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>