issue #1674 - workaround for HL7 FHIR-13601
The built-in SearchParameter-clinical-patient search parameter should
have a single target of type "Patient" but instead it has a target of
both "Patient" and "Group".

In this commit, I move the fhir-registry IndexGenerator from
src/main/java to src/test/java and add a step to modify this one spec
resource so that users can search on the `patient` search parameter
without needing a type modifier.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>