Fix issue with operation ordering on clean database
All my testing has been done on a dirty database. There are automated
build failures that appear to be related to test order and a resource
not existing yet on a clean database. Flipping the order of the resource
puts and initial clear cache operations so that the resource is
guaranteed to be in place.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>