Adjust compartment search handling (again)
In the previous commit, I allowed all non-patient compartment searches
through.  However, this could allow a user to gain information about the
existence (or non-existence) of resources they shouldn't know about. In
this revision, patient access to the Device, Practitioner, and
RelatedPerson compartments is once again restricted.
Patient access to the Encounter compartment is scoped by whether or not
that encounter is in the patient compartment.

We continue to allow searching of "other" compartments (i.e. custom
compartment definitions configured by the server operator).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>