#2421 - Add terminology clear-cache operations (#2422)
* #2421 - Add terminology clear-cache operations

The terminology subsystem caches terminology resources by URL and/or URL
+ version. If the user tries to update a CodeSystem or ValueSet resource
keeping the same URL + version, the cache does not get cleared and some
operations (e.g. validate-code) do not return the expected results after
update. This change adds an operation ($clear-cache) that can be called
on either terminology resource and will clear the appropriate caches.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Address test failure due to unexpected test interactions.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Fix issue with operation ordering on clean database

All my testing has been done on a dirty database. There are automated
build failures that appear to be related to test order and a resource
not existing yet on a clean database. Flipping the order of the resource
puts and initial clear cache operations so that the resource is
guaranteed to be in place.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Address code format issues

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Split term cache operations into a separate module

Per discussion with the team, the cache operations are being split into
a separate module that will not be included in the default server build.
Since it isn't automatically included, the build automation scripts
needed to be updated.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Update POM version for new module

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Fix Windows integration, update README

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Apply suggested change

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Apply suggested change

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

Co-authored-by: Lee Surprenant <lmsurpre@us.ibm.com>

* change package name from javax.json to jakarta.json

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

Co-authored-by: Lee Surprenant <lmsurpre@us.ibm.com>