modify Provenance-patient parameter
Per https://jira.hl7.org/browse/FHIR-32876 this search parameter is
wrong in the latest version of the spec.  I thought it was wrong in our
packaged version (4.0.1) as well, but it turns out I was mistaken.

Still, I thought it would be good to keep this logic in place for when
we move to the R5 artifacts.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>