issue #2528 - change interpretation of lt and gt
For both date and number searches.

We use the range interpretation of the search prefixes. The spec says
`lt` means "the range below the search value intersects (i.e. overlaps)
with the range of the target value" and `gt` means "the range above the
search value intersects (i.e. overlaps) with the range of the target
value"

Previously, we interpreted that to mean that a search like `gt2016`
should return any value above `2016-01-01 00:00:00` in the server's
timezone.

However, this interpretation leads to confusing and unintuitive results.
Instead, we will now interpret the "search value" of 2016 to be the
range `[2016-01-01 00:00:00, 2017-01-01 00:00:00)` and therefor it will
only find target values above 2017-01-01 00:00:00 (inclusive).

And similarly for `lt`.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>