issue #2535 - wrap exceptions instead of losing them
In addition to their OperationOutcomeIssues, we now include the original
exception in the CausedBy of the FHIRRestBundledRequestException.

This way, if the FHIROperationException was unexpected and has no
corresponding issues, when it gets mapped to a 500 Server Error we will
automatically log the exception from
FHIRResource.exceptionResponse(FHIROperationException, Status).

Also changed "bundle request" to "bundled request" in
FHIRRestHelper.logBundleRequestCompletedMsg to match the language at the
start of the bundled request and to make it more clear its the
individual bundled entry that completed and not the entire bundle.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>