issue #2526 - deprecate the /Resource and /DomainResource endpoints
Because we use "Resource" as the resource type for whole-system search
and whole-system history interactions, it wasn't as simple as just
checking for abstract types in FHIRRestHelper.validateInteraction.
Instead, I now check that the type in the request path is valid directly
from the JAX-RS endpoints / bundle entry processing.
This has the additional benefit of catching URL path errors up front,
avoiding almost all work associated with a bad request.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>