Revert the change to FHIRResource.exceptionResponse
Instead of ensuring that every string field is encoded for HTML (the
previous approach for this PR), I reverted back to the original code
which only encodes OperationOutcome.Issue.diagnostics.

The plan is to more selectively encode specific fields that are known to
contain user-provided content (#2596). Once that is done, I think we
should remove the logic that encodes all of the diagnostic strings.

Additionally, since we no longer encode the whole thing, I added
Encode.forHtml back to the buildUnsupportedResourceTypeException helpers
(but just around the user-provided portion of the error message).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>