issue #2184 - add derby-specific readCommonTokenValueId and do cleanup
* the derby-specific readCommonTokenValueId implementation replaces the
`(VALUES (?,value1), (?,value2), ...` construct with a temporary
table--following the same pattern used for upsertCommonTokenValues.
* removed dead code from getReferenceFilter; we had code to handle query
parameters of type reference which had modifiers that we don't actually
support for reference parameters
* removed commented-out code in from populateCodesSubSegment; I
convinced myself that we can use the common_token_values table for both
the IN and NOT-IN cases.
* removed stale getCommonTokenValueIds code / alternate implementations
* reordered field names and select statements to follow the convention
of existing methods (value, system, id)--personally I prefer
(system,code) to (code,system) but I opted for consistency with the
current codebase
* added defensive code to return an empty set of CommonTokenValueResult
if passed an empty Collection of CommonTokenValue

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>