issue #2266 - use system properties to create the right factories
In cases where fhir-model is used outside the server, many programs will
need to set the javax.xml.stream properties to get the right behavior.

This commit makes it so that, if these properties are NOT set, then
we'll temporarilly set them to get the right factory for our model.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>