issue #2742 - aggregate exit status of all executions in bulkdata-status
I found that the liberty implementation of java batch has a "feature"
where it only restarts partitions that have *not* completed
successfully.
Unfortunately, I couldn't find a simple way to get the ExitStatus of
successful partitions from previous executions and stuff those into the
ExitStatus of the latest execution.
Therefore, to get a complete list of partition results, it is necessary
for the status endpoint to aggregate the Exit Status of *all* of the
jobs executions.
Previously we were only checking the last execution and that was leading
to confusing behavior where the status shows partial results from the
export (even though the entire job was completed).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>