fhir-smart fixes
1. Add null check in afterRead and afterVRead.  In the case of
`afterRead`, this happens when users submit an update for a resource
that doesn't exist yet. We missed til now because our fhir-smart usage
was all read-only.

2. Stop trying to decode the signature of the JWT. Per the JWT spec,
this part isn't actually Base64 encoded and so the existing code will
only work for JWTs with an empty signature (which just happens to be
what I tested). I added a new test with the expanded coverage.

3. Update Claim.asString and Claim.asList to match the java-jwt behavior
of these methods, add javadoc to this effect, and assert in tests.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>