Condense packaged JSON resources during the build phase
In the past, we kept these resources in their condensed / non-pretty
format, but that made it hard to review changes and so we recently
converted them to their pretty-printed flavor instead.

The problem there is that the extra whitespace leads to a noticable size
difference for the jars.

On my system, removing this excess whitespace during the build
reduces the size between 10 and 18%:

carin-bb:          933,281 to 766,640
us-core:         1,443,884 to 1,212,573
fhir-registry:  19,585,400 to 17,758,332

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>