Condense packaged JSON resource during the build phase via profile
In https://github.com/IBM/FHIR/pull/2860 I added the condense-json action to the build section
of each project where we want it.

The problem with that is it requires everyone to have a built copy of the fhir-tools project
in order for our fhir-parent build to be successful.

This PR updates the approach by introducing a build profile in fhir-parent (and fhir-examples)
for controlling whether to strip the formatting of **all** JSON resources across all modules.

Additionally, this PR adds this new profile to a select few of the CI actions *and* our release action
(so that the release jars will benefit from the savings).

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>