Introduce fhir-server-spi
Currently, this module houses the SPI-related classes for:
* Interceptors
* Custom Operations

As part of that effort, I moved the Interaction enum from FHIRRestHelper
to its interface, FHIRResourceHelpers.

I updated fhir-smart to depend on fhir-server-spi instead of
fhir-server.

I also updated all fhir-operation modules to depend on a
'provided'-scope fhir-server-spi instead of fhir-server, with the
following exceptions:
1. fhir-operation-erase because that one still depends on
RestAuditLogger which I decided not to move into the spi module
2. fhir-operation-term-cache because that depends on
ServerRegistryResourceProvider (to clear its internal cache)

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>