Add dummy expressions to enable more test cases
Previously I had commented out the runCodeTest and runSystemTest cases
because of limitations with our ability to generate SQL fragments, but
then Robin had the bright idea of inserting a dummy expressions so that
we can enable these test cases.

I also updated one test case, testAddSystemIfPresentWithNonNullCache, so
that it retrieves the expected code system id from the mock cache...I
assume that was the original intention of the test given its name.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>