performance improvement for ModelSupport.getElementInfoMap
While investigating startup performance, I found that ModelSupport takes
a long time (~2 seconds) to initialize.
The most obvious spot for improvement was ModelSupport.getElementInfoMap
which is:
1. reading a file that contains class names for all model classes
2. finding those using Class.forName
3. iterating this list twice, and
4. reflecting over each class and its parents (even if they've already
been visited in a previous iteration)

I addressed much of that, but we're still reflecting over hundreds of
classes and so its still not fast. In my local testing, the method is
about twice as fast now, but it still takes over a second to initialize.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>