add debug output and attempt to fix the CI
I recently introduced
ParametersSearchUtilTest.testVersionedSearchParameterFilter and this was
following the existing pattern of this test to set a new threadlocal
FHIRRequestContext.
To try cleaning this up a bit, I've added an @AfterMethod cleanup()
method that should restore the FHIRRequestContext back to the "default"
tenant after each test...hopefully on the same thread that executed the
tests.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>