Update ParametersMap to store disambiguated results
Previously, ParametersMap had to store all the built-in parameters and
make those available to SearchUtil for filtering/disambiguation.

Now that we apply filtering before populating ParametersMap, we can
select a "winner" in cases where there are multiple parameter with the
same code (or multiple parameters with the same url / canonical).

For now, I chose a "last insert wins" approach...which means that the
order we load the map in will matter.

Note: this change uncovered an issue in our fhir-persistence search
tests where we defined a parameter with code "code" that conflicts with
a base spec parameter. The fix was to disambiguate via
fhir-server-config.

Signed-off-by: Lee Surprenant <lmsurpre@us.ibm.com>